refactor: separate dev tool from dispatcher
